# docs/contracts/api/openapi.yaml
openapi: 3.1.0
info:
  title: Cluster Platform API
  version: 1.1.0
  license: { name: Proprietary, url: "https://cluster.example/legal" }
  description: Authenticated, authorization-filtered API. IDs are lowercase RFC 9562 UUIDv7 values and timestamps are RFC 3339 UTC values ending in Z.
servers:
  - url: https://api.cluster.example/api/v1
security:
  - bearerAuth: []
paths:
...
  /authorization/access-decisions:
    parameters: [{ $ref: "#/components/parameters/CorrelationId" }]
    post:
      {
        summary: Evaluate access from actor context and current AuthorizationRecordFacts,
        operationId: decideAccess,
        requestBody:
          {
            required: true,
            content:
              {
                application/json:
                  {
                    schema:
                      { $ref: "#/components/schemas/AccessDecisionRequest" },
                  },
              },
          },
        responses:
          {
            "200": { $ref: "#/components/responses/AccessDecision" },
            "400": { $ref: "#/components/responses/BadRequest" },
            "401": { $ref: "#/components/responses/Unauthorized" },
            "403": { $ref: "#/components/responses/Forbidden" },
            "404": { $ref: "#/components/responses/NotFound" },
            "409": { $ref: "#/components/responses/Conflict" },
          },
      }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    identitySession:
      type: apiKey
      in: cookie
      name: cluster_identity_session
      description: HttpOnly, Secure identity-session cookie issued by `POST /identity/login`.
    documentsWorkerToken:
      type: apiKey
      in: header
      name: X-Documents-Worker-Token
      description: Internal document worker credential. It is never issued by this API.
  headers:
    Correlation:
      description: Required request correlation value echoed by the service; lowercase RFC 9562 UUIDv7.
      required: true
      schema: { $ref: "#/components/schemas/UUIDv7" }
...
  parameters:
    CorrelationId:
      {
        name: X-Correlation-ID,
        in: header,
        required: true,
        description: "Lowercase RFC 9562 UUIDv7 used for request, response, audit, and event correlation.",
        schema: { $ref: "#/components/schemas/UUIDv7" },
      }
...
  responses:
...
    BadRequest:
      {
        description: Invalid request (RFC 7807),
        headers:
          { X-Correlation-ID: { $ref: "#/components/headers/Correlation" } },
        content:
          {
            application/problem+json:
              { schema: { $ref: "../schemas/problem-details.schema.json" } },
          },
      }
    Unauthorized:
      {
        description: Authentication required or invalid (RFC 7807),
        headers:
          { X-Correlation-ID: { $ref: "#/components/headers/Correlation" } },
        content:
          {
            application/problem+json:
              { schema: { $ref: "../schemas/problem-details.schema.json" } },
          },
      }
...
    Forbidden:
      {
        description: Authorization denied (RFC 7807),
        headers:
          { X-Correlation-ID: { $ref: "#/components/headers/Correlation" } },
        content:
          {
            application/problem+json:
              { schema: { $ref: "../schemas/problem-details.schema.json" } },
          },
      }
    NotFound:
      {
        description: Resource not found or intentionally concealed (RFC 7807),
        headers:
          { X-Correlation-ID: { $ref: "#/components/headers/Correlation" } },
        content:
          {
            application/problem+json:
              { schema: { $ref: "../schemas/problem-details.schema.json" } },
          },
      }
    Conflict:
      {
        description: State or idempotency conflict (RFC 7807),
        headers:
          { X-Correlation-ID: { $ref: "#/components/headers/Correlation" } },
        content:
          {
            application/problem+json:
              { schema: { $ref: "../schemas/problem-details.schema.json" } },
          },
      }
...
    AccessDecision:
      {
        description: Explainable access decision,
        headers:
          { X-Correlation-ID: { $ref: "#/components/headers/Correlation" } },
        content:
          {
            application/json:
              { schema: { $ref: "../schemas/access-decision.schema.json" } },
          },
      }
...
  schemas:
...
    UUIDv7:
      {
        type: string,
        format: uuid,
        pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
        description: Lowercase RFC 9562 UUID version 7,
      }
...
    AccessDecisionRequest:
      {
        type: object,
        additionalProperties: false,
        required: [action, access_context, record_facts],
        properties:
          {
            action: { type: string, minLength: 1, maxLength: 128 },
            access_context: { $ref: "../schemas/access-context.schema.json" },
            record_facts: { $ref: "../schemas/record-facts.schema.json" },
          },
      }
...